+++ packages/plugin-zod/src/types.ts
@@ -1,5 +1,5 @@
 import type { Group, Output, PluginFactoryOptions, ResolveNameParams } from '@kubb/core'
-import type { Oas, SchemaObject } from '@kubb/oas'
+import type { Oas, SchemaObject, contentType } from '@kubb/oas'
 import type { Exclude, Generator, Include, Override, ResolvePathOptions, Schema } from '@kubb/plugin-oas'
 
 export type Options = {
@@ -7,6 +7,11 @@ export type Options = {
    * @default 'zod'
    */
   output?: Output<Oas>
+  /**
+   * Define which contentType should be used.
+   * By default, the first JSON valid mediaType will be used
+   */
+  contentType?: contentType
   /**
    * Group the Zod schemas based on the provided name.
    */